# Benchmark SQLite performance against prepared statements