I was benchmarking db to see if it could replace some "raw" SQLite code that I have, and I found that your existing benchmarks don't reflect optimal SQLite usage because they don't use prepared statements (nor does any part of the code seem to do that). I added them to the benchmarks, and got the following results. Before:

```
BenchmarkAppendRawSQL	     100	  14886563 ns/op
BenchmarkAppendUpper	     100	  15604206 ns/op
BenchmarkAppendTxRawSQL	  200000	      6672 ns/op
BenchmarkAppendTxUpper	  100000	     22329 ns/op
BenchmarkAppendTxUpperMap	   50000	     24342 ns/op
```

After:

```
BenchmarkAppendRawSQL	     100	  13728325 ns/op
BenchmarkAppendUpper	     100	  15739658 ns/op
BenchmarkAppendTxRawSQL	  500000	      3447 ns/op
BenchmarkAppendTxUpper	  100000	     22208 ns/op
BenchmarkAppendTxUpperMap	   50000	     26936 ns/op
```

This shows that the performance gap between db and raw SQL is actually a factor two worse than the original benchmarks showed...