@xiam 2015-06-11
Thanks a lot @larsmans. @pkieltyka and I have been discussing this speed problem too, he also mentioned using prepared statements. This is actually one of our biggest concerns to date.