@xiam 2015-06-11
BTW, we're going to track speed issues, improvements and benchmarks here: https://github.com/upper/db/issues/84.