@pkieltyka 2015-06-11
also, @larsmans which branch of upper/db are you using? you should be trying this on the v2 branch