@larsmans 2015-06-11
I'm on master. I also have a partial solution that caches the last query as a prepared statement, should I switch branches?